Refactor Project overlays into separate components
Moved ProjectOverlay, InputMap, InputTable, ScenarioAlert, and ShowHideCardsButton into their own files to improve modularity and maintainability.
